Add use_krpc flag for the DataStreamSender/Recvr/Mgr service
This patch adds a flag that switches between KRPC and Thrift for the
DataStream* service, or in other words the TransmitData RPC family.

The flag defaults to false, which means that Thrift will be used by
default.

This patch reverts the runtime filter set of RPCs to use Thrift by
default and always.

Since this patch is built on top of other patches that have configured
the Statestore and the StatestoreSubscriber to use KRPC by default,
the flag has no effect on those services; i.e. if use_krpc=false,
the Statestore and the StatestoreSubscriber will still use KRPC.
This will be changed at a later point, since this patch is a POC
for now.

TODO:
- The Thrift version of the DataStream* classes are renamed to
  OldDataStream*. We should ideally leave them with their original names
  and rename the new KRPC DataStream* classes to NewDataStream* or
  KRPCDataStream*, etc.
- Add an abstract class that the DataStream* and NewDataStream* classes
  can inherit from instead of using if/else logic to start up the
  appropriate class.
- Some of the BE tests most likely won't compile, fix them.
- Port on top of ASF master.

Change-Id: Ic8db9a83dfd892773a9a35df05e516134e4b1ea0