Pull in krb5_is_config_principal() when running against older kerberos versions
On kerberos versions < krb5-1.8, the function krb5_is_config_principal()
does not exist. Since our code uses that function, and we statically link
against kerberos, we would be unable to run on systems that have these
old kerberos versions.

We detect at runtime if the function exists in the library that we linked
against, and accordingly compile the re-implemented function, which is
under a #if defined().

Testing: Tested that it works on platforms with older and newer kerberos
versions ( < 1.8 and >= 1.8)

Change-Id: I3497de17741a559d83319e173090d0c29de4aa03