IMPALA-4671: (part-1) Copy kudu::ServicePool into the Impala namespace
The KuduRPC subsystem uses kudu::ServicePool to service all incoming
RPCs. Since this lives inside the Kudu codebase, all instrumentation
of RPCs flowing through the KuduRPC subsystem is not visible to Impala,
as Kudu does not export this instrumentation, but only maintains it
internally within kudu::ServicePool.

In order to reliably view the instrumentation of all RPCs flowing through
KRPC, one option is to modify kudu::ServicePool to take their
instrumentation and display it on our webpages, which is very invasive.

A second option is to have a parallel implementation of kudu::ServicePool
which for all purposes behaves the same way, but instead has this extra
code that displays this instrumentation in our webpages.

This patch is Part 1 of the second option. In this patch, we simply copy
the kudu::ServicePool code into be/src/rpc/ and rename it to ImpalaServicePool,
and do the necessary naming adjustments to make sure it compiles. This new
class is not used in this patch. It will be used in a follow on patch that
replaces the use of kudu::ServicePool with ImpalaServicePool by our RpcMgr.

We split it up this way so that it will be easy to review.

Change-Id: Ia3e0fa9ae22c4803b0cea736eb35e333e2671448