Improve Kerberos handshake performance by using in-memory replay cache
The Kerberos replay cache is in place to protect against replay attacks.
This cache however, is terribly inefficient as it writes to disk on
every request serviced. There is an option to turn off the cache, however,
that would leave the application open to attacks.

The next best option is to set the replay cache directory to the tmpfs,
or /dev/shm, which resides completely in RAM.

This patch checks for the existance of /dev/shm and sets the replay
cache directory to that if found. We expect connection setup speeds
on kerberized setups to improve greatly after this change.

Change-Id: Ic5c089dd94e3610da6f2ad1c76429ae0bed563ec